backend/requirements.txt
dj-database-url
Django==2.2.*
django-allauth
Faker
gunicorn
Pillow
psycopg2
pytz
sqlparse
whitenoise
djangorestframework
django-cors-headers
dj-rest-auth
pyyaml
uritemplate
drf-yasg
